Working Init Script gh-6
I have officially gotten the init script to work.

Apparently, my issue, as we discussed, was as simple as a bad directory call.
This code executes a led sweep sequence on start-up ("EAS-RV12-Startup-Leds.sh") and then executes "Enable_third_i2c.sh"

When shutting down, this code will flash all leds simultaneously for a while ("EAS-RV12-Shutdown-Leds.sh")

The Led flashes are only used as debug code and are not inherently necessary.

However, something you may be interested in knowing:
	THE SCRIPT WILL RUN WITHOUT LOGIN!

This code takes about 4 minutes after boot to run because of the inherent slowness of the "find" command and the fact that I require the script to start after everything else (init wise) has been executed.