Update EASEagle.lbr
Added two new symbols. RJ454Xsig is the name of the pinout for the analog cape that carries signals to the adc. RJ45I2C2XPWM is a temporary placeholder for the connection that is being used for the pitot tube, where the final pinout for pwm cap is not finalised.

Created a new Device called RJ454XSIG which connects the symbol of the same name to the footprint RJ45-8.